requirements.txt
asgiref==3.8.1
azure-core==1.30.1
azure-core-tracing-opentelemetry==1.0.0b11
azure-monitor-opentelemetry==1.4.0
azure-monitor-opentelemetry-exporter==1.0.0b24
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
fixedint==0.1.6
Flask==3.0.3
gunicorn==22.0.0
idna==3.7
importlib-metadata==7.0.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
msrest==0.7.1
oauthlib==3.2.2
opentelemetry-api==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-dbapi==0.45b0
opentelemetry-instrumentation-django==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-flask==0.45b0
opentelemetry-instrumentation-psycopg2==0.45b0
opentelemetry-instrumentation-requests==0.45b0
opentelemetry-instrumentation-urllib==0.45b0
opentelemetry-instrumentation-urllib3==0.45b0
opentelemetry-instrumentation-wsgi==0.45b0
opentelemetry-resource-detector-azure==0.1.4
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
packaging==24.0
psutil==5.9.8
psycopg2-binary==2.9.9
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==69.5.1
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
zipp==3.18.1
